wip: grouping UI-related configs together
